Refactor MCP resources and prompts
